Updated model
